+++ Userland/Libraries/LibUnicode/CharacterTypes.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Tim Flynn <trflynn89@serenityos.org>
+ * Copyright (c) 2021-2023, Tim Flynn <trflynn89@serenityos.org>
  *
  * SPDX-License-Identifier: BSD-2-Clause
  */
@@ -63,10 +63,5 @@ bool code_point_has_block(u32 code_point, Block block);
 bool code_point_has_grapheme_break_property(u32 code_point, GraphemeBreakProperty property);
 bool code_point_has_word_break_property(u32 code_point, WordBreakProperty property);
 bool code_point_has_sentence_break_property(u32 code_point, SentenceBreakProperty property);
-
-Vector<size_t> find_grapheme_segmentation_boundaries(Utf16View const&);
-Vector<size_t> find_word_segmentation_boundaries(Utf8View const&);
-Vector<size_t> find_word_segmentation_boundaries(Utf16View const&);
-Vector<size_t> find_sentence_segmentation_boundaries(Utf16View const&);
 
 }
